+++ packages/support/cldr-data-downloader/src/available_locales.js
@@ -1,0 +1,46 @@
+/**
+ * Download tool for Unicode CLDR JSON data
+ *
+ * Copyright 2021 Microsoft Corporation
+ * Copyright 2013 Rafael Xavier de Souza
+ *
+ * Released under the MIT license
+ */
+
+'use strict';
+
+const fs = require('fs');
+const path = require('path');
+
+function AvailableLocales(destPath) {
+  const mainPath = path.join(destPath, 'main');
+
+  this.availableLocales = fs.readdirSync(mainPath).filter(filepath => {
+    const stats = fs.statSync(path.join(mainPath, filepath));
+
+    return stats.isDirectory();
+  });
+
+  this.destPath = destPath;
+}
+
+const proto = AvailableLocales.prototype;
+
+proto.filepath = function () {
+  return path.join(this.destPath, 'availableLocales.json');
+};
+
+proto.toJson = function () {
+  return {
+    availableLocales: this.availableLocales
+  };
+};
+
+proto.write = function () {
+  // eslint-disable-next-line no-magic-numbers
+  const data = JSON.stringify(this.toJson(), null, 2);
+
+  fs.writeFileSync(this.filepath(), data);
+};
+
+module.exports = AvailableLocales;